Using utils for common functions
